chore: Copies returned by Jobs class getter
The method getJobs() in Jobs class (loaders package) used to return the
reference on the linked list of jobs. This provoked problems when using
algorithms that needed to modified the imported list of jobs. Instead of
reloading everything with the Test class for each scheduling algorithm I
chose to change the getter so that it returns a copy of the list. In
order to not be confused with getters that return references I renamed
the getJobs() method into copyJobs().