fix: Incorrect management of arrival dates in FIFO and RR
In FIFO and Round Robin (RR) if a task arrived with an arrival date
superior to the current date in the schedule the task would still be
marked as starting at current date in the schedule (e.g. we are at date
3 in the schedule, task is supposed to arrive at date 5 and it was
marked as arriving at date 3 anyway). Another problem to fix now: in EDF
tasks are sorted by deadline but arrival dates are not dealt with
properly...