fix: EDF algorithm (no quantum needed)
Quantum isn't a kind of critical section at the end of which you check
whether or not new jobs have arrived. At each job arrival/end you check
whether or not new jobs have arrived and reevaluate jobs order as a
consequence. To deal with this kind of preemption instead of a time step
method I chose to implement an event method, ie. at each schedule step I
check the next arrival date and and compute the end date of my schedule
entry as a consequence instead of checking at each unit of work if a new
job arrived.