fix: RMS algorithm (no quantum needed)
Similarly to EDF no quantum is truly needed as it is not a critical
section and priorities must be reevaluated at each job arrival/end. In
addition I also chose to reset period if a task is aperiodic so that it
has the lowest priority in the RMS algorithm. To do this I chose an
infinite period defined after job is loaded in Jobs.java.