feat: Added Metrics (first version)
This version of the Metrics class allows to compute the max tardiness,
the number of deadline misses, the IDs of late jobs and the total
makespan. For this I had to modify the ScheduleEntry definition. Now the
whole Job and the whole Server is passed to the constructor of the
ScheduleEntry. This had a positive effect on a lot of files where
getters aren't necessary anymore.