refactor: Removed unecessary methods in Scheduler and easier clock mgmt
Removed methods print and write from Scheduler as they only redirected
towards Schedule methods and didn't made sense regarding the role of the
Scheduler class. Scheduler now is more or less a servers manager
abstract class. Also, clock modifications (schedule.currentDate) are
only done inside runScheduleStep() redefinitions. This is an important
change as it's easier to guarantee the correctness of the schedule
currentDate and it also simplifies its maintainability.