chore: Added EnergyAware simulator and necessary classes
Added EnergyAware class in simulator package to run simulations for
energy aware FIFO and EDF. I also deactivated MonoServer and MultiServer
from Main as they won't be modified during the whole development of this
exercise. Finally I created two new classes: FIFOe and EDFe for energy
aware FIFO and EDF. This choice is based on the fact that unlike
multiserver allows to simulate monoserver with the same piece of code,
energy aware will deeply modify the behavior of scheduling in any case.