fix: Fixed late servers evaluation to all assigned jobs
Late servers evaluation used to deal with only the currently running
job. Now all jobs in the assignedJobs list of the server are considered.
Server frequencies are hence increased if at least one job is going to
be late. Surprisingly results in terms of consumption and deadline
misses are simultaneously improved!